% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-trial.R
\name{allocationRatio}
\alias{allocationRatio}
\alias{allocationRatio,Trial-method}
\title{Allocation Ratio}
\usage{
allocationRatio(x)

\S4method{allocationRatio}{Trial}(x)
}
\arguments{
\item{x}{A \link{Trial} object.}
}
\value{
Named integer vector of allocation weights.
}
\description{
Allocation Ratio
}
\examples{
allocationRatio(Trial("T1", c("active", "control")))
}
