% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-trial.R
\docType{class}
\name{Arm-class}
\alias{Arm-class}
\title{Trial Arm}
\description{
Trial Arm
}
\section{Slots}{

\describe{
\item{\code{label}}{Arm label.}

\item{\code{ratio}}{Positive integer allocation weight.}

\item{\code{description}}{Optional description.}

\item{\code{is_control}}{Whether this is a control arm.}

\item{\code{treatment_code}}{Optional masked treatment code.}
}}

