% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-trial.R
\name{Arm}
\alias{Arm}
\title{Construct a Trial Arm}
\usage{
Arm(
  label,
  ratio = 1L,
  description = NA_character_,
  is_control = FALSE,
  treatment_code = NA_character_
)
}
\arguments{
\item{label}{Non-empty arm label.}

\item{ratio}{Positive integer allocation weight.}

\item{description}{Optional description.}

\item{is_control}{Whether this is a control arm.}

\item{treatment_code}{Optional masked treatment code.}
}
\value{
An \link{Arm} object.
}
\description{
Construct a Trial Arm
}
\examples{
Arm("active", ratio = 2L, is_control = FALSE)
}
