% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-trial.R
\name{arms}
\alias{arms}
\alias{arms,Trial-method}
\title{Arm Labels}
\usage{
arms(x)

\S4method{arms}{Trial}(x)
}
\arguments{
\item{x}{A \link{Trial} object.}
}
\value{
Character vector of arm labels in allocation order.
}
\description{
Arm Labels
}
\examples{
arms(Trial("T1", c("active", "control")))
}
