% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{assignments}
\alias{assignments}
\alias{assignments,RandomizationSequence-method}
\title{Revealed Assignments}
\usage{
assignments(seq)

\S4method{assignments}{RandomizationSequence}(seq)
}
\arguments{
\item{seq}{A \linkS4class{RandomizationSequence} object.}
}
\value{
Assignment table with every unrevealed arm replaced by \code{NA}.
}
\description{
Revealed Assignments
}
\examples{
sequence <- randomize(Trial("T1", c("A", "B")), n = 4, seed = 1)
assignments(sequence)
}
