% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioTrial-package.R
\docType{package}
\name{PhysioTrial-package}
\alias{PhysioTrial}
\alias{PhysioTrial-package}
\title{PhysioTrial: Auditable Clinical-Trial Design and Analysis Workflows}
\description{
PhysioTrial provides trial data models, concealed allocation sequences,
treatment-code management, CONSORT/SPIRIT operations, adverse-event
summaries, explicit analysis populations, endpoint-analysis helpers, and
auditable CDISC SDTM-shaped and ADaM-shaped exports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/x-biosignal/PhysioTrial}
  \item \url{https://x-biosignal.r-universe.dev/PhysioTrial}
  \item \url{https://x-biosignal.github.io/PhysioTrial/}
  \item Report bugs at \url{https://github.com/x-biosignal/PhysioTrial/issues}
}

}
\author{
\strong{Maintainer}: Yusuke Matsui \email{mail.to.matsui@gmail.com}

}
\keyword{internal}
