% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{powerBinary}
\alias{powerBinary}
\title{Power of a two-arm binary endpoint at a given sample size}
\usage{
powerBinary(n1, p1, p2, sig_level = 0.05, ratio = 1, sides = 2L)
}
\arguments{
\item{n1}{Control-arm sample size.}

\item{p1, p2, sig_level, ratio, sides}{As in \code{\link[=sampleSizeBinary]{sampleSizeBinary()}}.}
}
\value{
The achieved power.
}
\description{
Power of a two-arm binary endpoint at a given sample size
}
\seealso{
\code{\link[=sampleSizeBinary]{sampleSizeBinary()}}
}
