% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{randomSeed}
\alias{randomSeed}
\alias{randomSeed,RandomizationSequence-method}
\title{Captured Randomization Seed}
\usage{
randomSeed(seq)

\S4method{randomSeed}{RandomizationSequence}(seq)
}
\arguments{
\item{seq}{A \linkS4class{RandomizationSequence} object.}
}
\value{
Positive integer seed.
}
\description{
Captured Randomization Seed
}
\examples{
sequence <- randomize(Trial("T1", c("A", "B")), n = 4, seed = 19)
randomSeed(sequence)
}
