% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-trial.R
\name{strata}
\alias{strata}
\alias{strata,Trial-method}
\title{Trial Strata}
\usage{
strata(x)

\S4method{strata}{Trial}(x)
}
\arguments{
\item{x}{A \link{Trial} object.}
}
\value{
Named list of prognostic factors and admissible levels.
}
\description{
Trial Strata
}
\examples{
strata(Trial(
  "T1", c("active", "control"),
  strata = list(site = c("north", "south"))
))
}
