% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R, R/blinding.R
\name{trialFingerprint}
\alias{trialFingerprint}
\alias{trialFingerprint,RandomizationSequence-method}
\alias{trialFingerprint,BlindingManager-method}
\title{Trial Object Fingerprint}
\usage{
trialFingerprint(x)

\S4method{trialFingerprint}{RandomizationSequence}(x)

\S4method{trialFingerprint}{BlindingManager}(x)
}
\arguments{
\item{x}{A \linkS4class{RandomizationSequence} or \linkS4class{BlindingManager} object.}
}
\value{
Deterministic \code{xxhash64} semantic-content fingerprint.
}
\description{
Trial Object Fingerprint
}
\examples{
sequence <- randomize(Trial("T1", c("A", "B")), n = 4, seed = 1)
trialFingerprint(sequence)
}
